requirements.txt
# Streamlit and related packages
streamlit==1.2.0

# Machine learning and data manipulation libraries
numpy>=1.21,<2.0
pandas>=1.3,<2.0
tensorflow>=2.8,<3.0
torch==1.10.0
torchvision==0.11.1
h5py>=3.1,<4.0  # Required for TensorFlow compatibility

# Web framework and server
fastapi==0.68.0
uvicorn==0.15.0

# Additional libraries for functionality
Pillow>=8.0,<10.0  # For image processing
requests>=2.25,<3.0  # For making HTTP requests
rich>=10.0,<13.0  # For rich text formatting in the console

# Other necessary libraries (add or remove as needed)
absl-py==2.1.0
anyio==3.6.2
click==8.1.7
colorama==0.4.6
filelock==3.16.1
flatbuffers==24.3.25
fsspec==2024.10.0
gast==0.6.0
google-pasta==0.2.0
grpcio==1.68.1
idna==3.10
Jinja2==3.1.4
keras==3.7.0  # If using Keras for your model, otherwise remove it.
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe>=2, <3  # Ensure compatibility with Jinja2.
ml-dtypes>=0, <1  # Specify a valid minimum version.
mpmath>=1, <2  # Specify a valid minimum version.
networkx>=3, <4  # Specify a valid version range.
opt_einsum>=3, <4  # Use the latest stable version.
packaging>=24, <25  # Use the latest stable version.
protobuf>=5, <6  # Use the latest stable version.
pydantic>=1, <2  # Use Pydantic v1 for compatibility with FastAPI.
six>=1, <2  # Use the latest stable version.
sniffio>=1, <2  # Use the latest stable version.
starlette>=0, <1  # Use the latest stable version compatible with FastAPI.
sympy>=1, <2  # Use the latest stable version.
tensorboard>=2, <3  # Use the latest stable version compatible with TensorFlow.
tensorflow-io-gcs-filesystem>=0, <1  # Use the latest stable version compatible with TensorFlow.
termcolor>=2, <3  # Use the latest stable version.
typing_extensions>=4, <5  # Use the latest stable version.
urllib3>=1, <2  # Use the latest stable version.
wrapt>=1, <2  # Use the latest stable version.

# Optional: If you have local package code to include, uncomment below:
# -e .  